backend/pyproject.toml
[tool]
[tool.poetry]
name = "undercard_18898"
description = "undercard django backend"
authors = []
version = "0.0.0"

[tool.poetry.dependencies]
python = "~3.7"
boto3 = "*"
celery = {version = "==5.0.5", extras = ["redis"]}
cryptography = "==3.3.*,>=3.3.1"
django = "==2.2.*,>=2.2.3"
django-allauth = "*"
django-autoslug = "*"
django-bootstrap4 = "==1.1.*,>=1.1.1"
django-celery-beat = "*"
django-cors-headers = "*"
django-debug-toolbar = "*"
django-environ = "==0.4.*,>=0.4.5"
django-extensions = "==2.2.*,>=2.2.9"
django-flexible-subscriptions = "*"
django-imagekit = "*"
django-redis = "*"
django-rest-auth = "==0.9.*,>=0.9.5"
django-storages = "*"
django-taggit = "==1.3.*,>=1.3.0"
djangorestframework = "==3.11.*,>=3.11.0"
drf-extra-fields = "*"
drf-tweaks = "*"
drf-yasg = "==1.17.*,>=1.17.1"
fcm-django = "==0.3.*,>=0.3.4"
ics = "*"
importlib-metadata = "*"
ipython = ">=7.17.0"
numba = "*"
numpy = "==1.19.5"
packaging = "*"
pandas = "==1.1.5"
pillow = "*"
# psycopg2 = "==2.8.*,>=2.8.5"
psycopg2-binary = "*"
pyjwt = "*"
python-benedict = "*"
python-dateutil = "*"
pyyaml = "==5.3.*,>=5.3.1"
sendgrid = "==6.6.0"
sentry-sdk = "*"
stripe = "*"
tatsu = "==4.4.0"
timezonefinder = "*"
traitlets = "==4.3.3"
twilio = "*"
waitress = "==1.4.*,>=1.4.3"
whitenoise = "==5.1.*,>=5.1.0"
django-push-notifications = {extras = ["WP", "APNS"], version = "^2.0.0"}
factory-boy = "^3.2.0"
ipdb = "^0.13.9"
django-polymorphic = "^3.0.0"
pylint = "^2.10.2"
